Fix bug: if user1/admin1 was on a project page containing for example image1, then another admin deleted image1, previously user1/admin1 would still be able to view the custom metadata of image1 and delete image1 (even though it had been deleted) (assuming user1/admin1 did not refresh the page).

After the fix: user1/admin1 would not be able to view custom metadata or delete image1.

+++ dotnet-backend/Core/Services/ProjectService.cs
@@ -501,7 +501,24 @@ namespace Core.Services
         
         public async Task DeleteAssetFromProject(int projectId, string blobId)
         {
-            await _repository.DeleteAssetFromProjectInDb(projectId, blobId);
+            // verify that asset hasn't already been deleted (e.g. by another user), if it has, then return InvalidOperationException
+            try {
+                var asset = await this.GetAssetObject(blobId);
+                if (asset == null) {
+                    throw new DataNotFoundException();
+                }
+                await _repository.DeleteAssetFromProjectInDb(projectId, blobId);
+            } catch (DataNotFoundException) {
+                throw;
+            } catch (Exception ex) {
+                throw;
+            }
+        }
+
+        public async Task<Asset> GetAssetObject(string blobId) { // if asset has been deleted, returns DataNotFoundException
+
+            return await _repository.GetAssetObject(blobId);
+
         }
 
         public async Task<string?> GetCustomMetadataNameByIdAsync(int fieldID)
